refactor BasicAction and RadAction into a single abstract class that each specific action overrides as needed
