+++ build.gradle.kts
@@ -115,6 +115,9 @@ tasks {
         version.set(remoteRobotVersion)
     }
 
+    runIde {
+        systemProperty("idea.is.internal", "true");
+    }
 
     // Configure UI tests plugin
     // Read more: https://github.com/JetBrains/intellij-ui-test-robot
