+++ src/main/java/network/radicle/jetbrains/radiclejetbrainsplugin/RadicleBundle.java
@@ -20,6 +20,10 @@ public class RadicleBundle extends DynamicBundle {
         return INSTANCE.messageOrDefault(key, key, params);
     }
 
+    public static @Nls @NotNull String message(@NotNull @NonNls String key, String defaultMessage, Object @NotNull ... params) {
+        return INSTANCE.messageOrDefault(key, defaultMessage, params);
+    }
+
     public static Supplier<String> lazyMessage(@NotNull @NonNls String key, Object @NotNull ... params) {
         return () -> message(key, params);
     }
