+++ src/main/java/network/radicle/jetbrains/radiclejetbrainsplugin/actions/RadiclePullAction.java
@@ -7,6 +7,7 @@ import com.intellij.openapi.project.Project;
 import git4idea.repo.GitRepositoryManager;
 import network.radicle.jetbrains.radiclejetbrainsplugin.RadicleBundle;
 import network.radicle.jetbrains.radiclejetbrainsplugin.UpdateBackgroundTask;
+import network.radicle.jetbrains.radiclejetbrainsplugin.actions.rad.RadAction;
 import network.radicle.jetbrains.radiclejetbrainsplugin.actions.rad.RadPull;
 import org.jetbrains.annotations.NotNull;
 
@@ -14,10 +15,6 @@ import java.util.concurrent.CountDownLatch;
 import java.util.concurrent.atomic.AtomicBoolean;
 
 public class RadiclePullAction extends AnAction {
-
-    protected CountDownLatch updateCountDown;
-    protected AtomicBoolean executingFlag = new AtomicBoolean(false);
-
     @Override
     public void actionPerformed(@NotNull AnActionEvent e) {
         var project = e.getProject();
@@ -27,21 +24,21 @@ public class RadiclePullAction extends AnAction {
     public void performAction(Project project) {
         var gitRepoManager = GitRepositoryManager.getInstance(project);
         var repos = gitRepoManager.getRepositories();
-        if (!BasicAction.isCliPathConfigured(project)) {
+        if (!RadAction.isCliPathConfigured(project)) {
             return ;
         }
         ApplicationManager.getApplication().executeOnPooledThread(() -> {
-            var radInitializedRepos = BasicAction.getInitializedReposWithNodeConfigured(repos, true);
+            var radInitializedRepos = RadAction.getInitializedReposWithNodeConfigured(repos, true);
             if (radInitializedRepos.isEmpty()) {
                 return ;
             }
-            updateCountDown = new CountDownLatch(radInitializedRepos.size());
+            final var updateCountDown = new CountDownLatch(radInitializedRepos.size());
             radInitializedRepos.forEach(repo -> ApplicationManager.getApplication().executeOnPooledThread(() -> {
                 var pull = new RadPull(repo);
-                new BasicAction(pull, repo.getProject(), updateCountDown).perform();
+                pull.perform(updateCountDown);
             }));
             UpdateBackgroundTask ubt = new UpdateBackgroundTask(project, RadicleBundle.message("radPullProgressTitle"),
-                    updateCountDown, executingFlag);
+                    updateCountDown);
             new Thread(ubt::queue).start();
         });
     }
