+++ src/main/java/network/radicle/jetbrains/radiclejetbrainsplugin/actions/RadicleSyncAction.java
@@ -8,17 +8,15 @@ import git4idea.repo.GitRepository;
 import git4idea.repo.GitRepositoryManager;
 import network.radicle.jetbrains.radiclejetbrainsplugin.RadicleBundle;
 import network.radicle.jetbrains.radiclejetbrainsplugin.UpdateBackgroundTask;
+import network.radicle.jetbrains.radiclejetbrainsplugin.actions.rad.RadAction;
 import network.radicle.jetbrains.radiclejetbrainsplugin.actions.rad.RadSync;
 import org.jetbrains.annotations.NotNull;
 
 import java.util.List;
 import java.util.concurrent.CountDownLatch;
-import java.util.concurrent.atomic.AtomicBoolean;
 
 public class RadicleSyncAction extends AnAction {
-
     protected CountDownLatch updateCountDown;
-    protected AtomicBoolean executingFlag = new AtomicBoolean(false);
 
     @Override
     public void actionPerformed(@NotNull AnActionEvent e) {
@@ -34,21 +32,21 @@ public class RadicleSyncAction extends AnAction {
     }
 
     public void performAction(@NotNull Project project, @NotNull List<GitRepository> repos) {
-        if (!BasicAction.isCliPathConfigured(project)) {
+        if (!RadAction.isCliPathConfigured(project)) {
             return ;
         }
         ApplicationManager.getApplication().executeOnPooledThread(() -> {
-            var radInitializedRepos = BasicAction.getInitializedReposWithNodeConfigured(repos, true);
+            var radInitializedRepos = RadAction.getInitializedReposWithNodeConfigured(repos, true);
             if (radInitializedRepos.isEmpty()) {
                 return;
             }
             updateCountDown = new CountDownLatch(radInitializedRepos.size());
             radInitializedRepos.forEach(repo -> ApplicationManager.getApplication().executeOnPooledThread(() -> {
                 var sync = new RadSync(repo);
-                new BasicAction(sync, repo.getProject(), updateCountDown).perform();
+                sync.perform(updateCountDown);
             }));
             UpdateBackgroundTask ubt = new UpdateBackgroundTask(project, RadicleBundle.message("radSyncProgressTitle"),
-                    updateCountDown, executingFlag);
+                    updateCountDown);
             new Thread(ubt::queue).start();
         });
     }
