+++ src/main/java/network/radicle/jetbrains/radiclejetbrainsplugin/actions/rad/RadAuth.java
@@ -3,18 +3,11 @@ package network.radicle.jetbrains.radiclejetbrainsplugin.actions.rad;
 import com.google.common.base.Strings;
 import com.intellij.execution.process.ProcessOutput;
 import com.intellij.openapi.application.ApplicationManager;
-import git4idea.repo.GitRepository;
-import network.radicle.jetbrains.radiclejetbrainsplugin.RadicleBundle;
-import network.radicle.jetbrains.radiclejetbrainsplugin.actions.BasicAction;
 import network.radicle.jetbrains.radiclejetbrainsplugin.services.RadicleApplicationService;
 
 import java.util.List;
 
-public class RadAuth implements RadAction {
-
-    private String notificationSuccessMsg;
-    private String errorMsg;
-    private String successMsg;
+public class RadAuth extends RadAction {
     private final String name;
     private final String passphrase;
     private final RadAuthAction action;
@@ -23,31 +16,36 @@ public class RadAuth implements RadAction {
         this.name = name;
         this.passphrase = passphrase;
         this.action = action;
-        setMessages();
     }
 
-    private void setMessages() {
+    @Override
+    public String getActionName() {
         if (action == RadAuthAction.CREATE_IDENTITY) {
-            successMsg = RadicleBundle.message("createIdentitySuccess");
-            notificationSuccessMsg = RadicleBundle.message("createIdentitySuccess");
-            errorMsg = RadicleBundle.message("createIdentityError");
+            return "AuthCreateIdentity";
         } else if (action == RadAuthAction.SET_DEFAULT_IDENTITY) {
-            successMsg = RadicleBundle.message("setDefaultIdentitySuccess");
-            notificationSuccessMsg =  RadicleBundle.message("setDefaultIdentitySuccess");
-            errorMsg = RadicleBundle.message("setDefaultIdentityError");
+            return "AuthSetDefaultIdentity";
         } else if (action == RadAuthAction.REMOVE_IDENTITY) {
-            successMsg = RadicleBundle.message("removeIdentitySuccess");
-            notificationSuccessMsg = RadicleBundle.message("removeIdentitySuccess");
-            errorMsg = RadicleBundle.message("removeIdentityError");
+            return "AuthRemoveIdentity";
+        }
+        return "";
+    }
+
+    @Override
+    public ProcessOutput run() {
+        switch (action) {
+            case CREATE_IDENTITY:
+                return createNewIdentity();
+            case REMOVE_IDENTITY:
+                return removeIdentity();
+            case SET_DEFAULT_IDENTITY:
+                return setDefaultIdentity();
+            default:
+                return new ProcessOutput(-1);
         }
     }
 
-    public enum RadAuthAction {
-        CREATE_IDENTITY, REMOVE_IDENTITY, SET_DEFAULT_IDENTITY
-    }
-
     public ProcessOutput removeIdentity() {
-        var storagePath = BasicAction.getStoragePath();
+        var storagePath = getStoragePath();
         if (Strings.isNullOrEmpty(storagePath.keysStoragePath) || Strings.isNullOrEmpty(storagePath.gitStoragePath)) {
             return new ProcessOutput(-1);
         }
@@ -67,37 +65,7 @@ public class RadAuth implements RadAction {
         return rad.auth(name,passphrase,false);
     }
 
-    @Override
-    public ProcessOutput run() {
-        switch (action) {
-            case CREATE_IDENTITY:
-                return createNewIdentity();
-            case REMOVE_IDENTITY:
-                return removeIdentity();
-            case SET_DEFAULT_IDENTITY:
-                return setDefaultIdentity();
-            default:
-                return new ProcessOutput(-1);
-        }
-    }
-
-    @Override
-    public String getErrorMessage() {
-        return errorMsg;
-    }
-
-    @Override
-    public String getSuccessMessage() {
-        return successMsg;
-    }
-
-    @Override
-    public String getNotificationSuccessMessage() {
-        return notificationSuccessMsg;
-    }
-
-    @Override
-    public GitRepository getRepo() {
-        return null;
+    public enum RadAuthAction {
+        CREATE_IDENTITY, REMOVE_IDENTITY, SET_DEFAULT_IDENTITY
     }
 }
