+++ src/main/java/network/radicle/jetbrains/radiclejetbrainsplugin/actions/rad/RadSync.java
@@ -6,38 +6,19 @@ import git4idea.repo.GitRepository;
 import network.radicle.jetbrains.radiclejetbrainsplugin.RadicleBundle;
 import network.radicle.jetbrains.radiclejetbrainsplugin.services.RadicleApplicationService;
 
-public class RadSync implements RadAction {
-
-    private GitRepository repo;
-
+public class RadSync extends RadAction {
     public RadSync(GitRepository repo) {
-        this.repo = repo;
+        super(repo);
+    }
+
+    @Override
+    public String getActionName() {
+        return "Sync";
     }
 
     @Override
     public ProcessOutput run () {
         var rad = ApplicationManager.getApplication().getService(RadicleApplicationService.class);
         return rad.sync(repo);
-    }
-
-    @Override
-    public String getErrorMessage() {
-        return RadicleBundle.message("errorInRadSync");
-    }
-
-    @Override
-    public String getSuccessMessage() {
-        return RadicleBundle.message("successInRadSync");
-    }
-
-    @Override
-    public String getNotificationSuccessMessage() {
-        return RadicleBundle.message("radSyncNotification");
-    }
-
-
-    @Override
-    public GitRepository getRepo() {
-        return repo;
     }
 }
