+++ src/main/java/network/radicle/jetbrains/radiclejetbrainsplugin/listeners/RadicleManagerListener.java
@@ -6,15 +6,13 @@ import com.intellij.openapi.diagnostic.Logger;
 import com.intellij.openapi.project.Project;
 import com.intellij.openapi.project.ProjectManagerListener;
 import network.radicle.jetbrains.radiclejetbrainsplugin.RadicleBundle;
-import network.radicle.jetbrains.radiclejetbrainsplugin.actions.BasicAction;
+import network.radicle.jetbrains.radiclejetbrainsplugin.actions.rad.RadAction;
 import network.radicle.jetbrains.radiclejetbrainsplugin.config.RadicleSettingsHandler;
 import network.radicle.jetbrains.radiclejetbrainsplugin.services.RadicleProjectService;
 
 import java.util.List;
 
 public class RadicleManagerListener implements ProjectManagerListener {
-
-    private RadicleSettingsHandler radicleSettingsHandler;
     private static final Logger log = Logger.getInstance(RadicleManagerListener.class);
 
     @Override
@@ -25,11 +23,11 @@ public class RadicleManagerListener implements ProjectManagerListener {
     }
 
     private void showSuccessNotification(Project project) {
-        radicleSettingsHandler = new RadicleSettingsHandler();
+        var radicleSettingsHandler = new RadicleSettingsHandler();
         var settings = radicleSettingsHandler.loadSettings();
         if (Strings.isNullOrEmpty(settings.getPath())) {
-            BasicAction.showNotification(project, "radicle", "installedSuccessfully", NotificationType.INFORMATION,
-                    List.of(new BasicAction.ConfigureRadCliNotificationAction(project, RadicleBundle.lazyMessage("configure"))));
+            RadAction.showNotification(project, "radicle", "installedSuccessfully", NotificationType.INFORMATION,
+                    List.of(new RadAction.ConfigureRadCliNotificationAction(project, RadicleBundle.lazyMessage("configure"))));
         }
     }
 }
