+++ src/main/resources/messages/RadicleBundle.properties
@@ -12,15 +12,32 @@ radCliWarning=Radicle CLI warning
 radDetected=Rad Detected
 allowRad=It looks like you git pushed in a radicle-initialised project, rather than rad push. Perhaps you also want to rad sync?
 autoDetected=Auto-detected:
-errorInRadSync=Error in rad sync
-successInRadSync=Success in rad sync
-radSyncNotification=Project synced
-errorInRadPull=Error in rad pull
-successInRadPull=Success in rad pull
-radPullNotification=Remotes fetched
-errorInRadPush=Error in rad push
-successInRadPush=Success in rad push
-radPushNotification=Project published successfully
+
+radError_Clone=Clone Failed
+radSuccess_Clone=Success in rad clone
+radError_Init=Unable to initialize the project
+radSuccess_Init=The project initialized successful
+radError_Pull=Error in rad pull
+radSuccess_Pull=Success in rad pull
+radNotification_Pull=Remotes fetched
+radError_Sync=Error in rad sync
+radSuccess_Sync=Success in rad sync
+radNotification_Sync=Project synced
+radError_Push=Error in rad push
+radSuccess_Push=Success in rad push
+radNotification_Push=Project published successfully
+radSuccess_AuthSetDefaultIdentity=The default profile was changed
+radNotification_AuthSetDefaultIdentity=The default profile was changed
+radError_AuthSetDefaultIdentity=There was an error setting default profile
+radSuccess_AuthCreateIdentity=The new identity was created successfully
+radNotification_AuthCreateIdentity=The new identity was created successfully
+radError_AuthCreateIdentity=There was an error creating new identity
+radSuccess_AuthRemoveIdentity=The identity was removed successfully
+radNotification_AuthRemoveIdentity=The identity was removed successfully
+radError_AuthRemoveIdentity=There was an error removing identity
+radError_SelfActiveProfile=There was an error finding active profile
+radError_SelfGetProfiles=There was an error finding profiles
+
 radPullProgressTitle=Pull in progress
 radSyncProgressTitle=Sync in progress
 seedNodeMissing=Seed node is missing
@@ -34,14 +51,6 @@ invalidPort=Invalid Port
 seedNodeExists=The seed node already exists
 port=Port
 newIdentity=Create a new Radicle Identity (Peer ID)
-setDefaultIdentitySuccess=The default profile change successfully
-setDefaultIdentityError=There is an error setting default profile
-createIdentitySuccess=The new identity created successfully
-createIdentityError=There is an error creating new identity
-removeIdentitySuccess=The identity removed successfully
-removeIdentityError=There is an error removing identity
-activeProfileError=There is an error finding active profile
-profilesError=There is an error finding profiles
 peerId=Peer ID
 default=Default
 active=Active
@@ -60,13 +69,9 @@ httpRequestErrorDesc=Unable to fetch projects
 cloneFailed=Clone Failed
 directoryError=Unable to create destination directory
 destinationDoesntExist=Destination doesn't exist
-errorInRadClone=Error in rad clone
-successInRadClone=Success in rad clone
 selectSeedNode=Select seed node to view projects
 selectProject=Select project
 shareProject=Share Project on Radicle
-initError=Unable to initialize the project
-initSuccess=The project initialized successful
 publishProgressBar=Publish in progress
 clone=Clone
 urn=URN:
